Db type and Sql query collection moved to DatabaseClientDecorator
