+++ dd-java-agent/agent-bootstrap/src/main/java/datadog/trace/bootstrap/instrumentation/decorator/DatabaseClientDecorator.java
@@ -1,15 +1,20 @@
 package datadog.trace.bootstrap.instrumentation.decorator;
 
+import static datadog.trace.api.gateway.Events.EVENTS;
 import static datadog.trace.bootstrap.instrumentation.api.Tags.DB_TYPE;
 
 import datadog.trace.api.Config;
 import datadog.trace.api.cache.DDCache;
 import datadog.trace.api.cache.DDCaches;
+import datadog.trace.api.gateway.RequestContext;
+import datadog.trace.api.gateway.RequestContextSlot;
 import datadog.trace.api.naming.NamingSchema;
 import datadog.trace.api.naming.SpanNaming;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
+import datadog.trace.bootstrap.instrumentation.api.AgentTracer;
 import datadog.trace.bootstrap.instrumentation.api.Tags;
 import datadog.trace.bootstrap.instrumentation.api.UTF8BytesString;
+import java.util.function.BiConsumer;
 
 public abstract class DatabaseClientDecorator<CONNECTION> extends ClientDecorator {
   protected static class NamingEntry {
@@ -45,11 +50,20 @@ public abstract class DatabaseClientDecorator<CONNECTION> extends ClientDecorato
 
   protected abstract String dbType();
 
+  protected String dbType(CONNECTION connection) {
+    return dbType();
+  }
+
   protected abstract String dbUser(CONNECTION connection);
 
   protected abstract String dbInstance(CONNECTION connection);
 
   protected abstract CharSequence dbHostname(CONNECTION connection);
+
+  // Extract this to allow for easier testing
+  protected AgentTracer.TracerAPI tracer() {
+    return AgentTracer.get();
+  }
 
   /**
    * This should be called when the connection is being used, not when it's created.
@@ -77,6 +91,22 @@ public abstract class DatabaseClientDecorator<CONNECTION> extends ClientDecorato
           span.setServiceName(hostName.toString());
         }
       }
+
+      if (Config.get().getAppSecRaspEnabled()) {
+        BiConsumer<RequestContext, String> connectDbCallback =
+            tracer()
+                .getCallbackProvider(RequestContextSlot.APPSEC)
+                .getCallback(EVENTS.databaseConnection());
+        if (connectDbCallback != null) {
+          RequestContext ctx = span.getRequestContext();
+          if (ctx != null) {
+            String dbType = dbType(connection);
+            if (dbType != null) {
+              connectDbCallback.accept(ctx, dbType);
+            }
+          }
+        }
+      }
     }
     return span;
   }
@@ -102,6 +132,28 @@ public abstract class DatabaseClientDecorator<CONNECTION> extends ClientDecorato
 
   public AgentSpan onStatement(final AgentSpan span, final CharSequence statement) {
     span.setResourceName(statement);
+    return span;
+  }
+
+  /**
+   * The method used to provide raw sql to prevent SQL-injection attacks SQL query should never be
+   * exposed because it may contain sensitive data.
+   */
+  public AgentSpan onStatementRaw(AgentSpan span, String sql) {
+    if (Config.get().getAppSecRaspEnabled()) {
+      BiConsumer<RequestContext, String> sqlQueryCallback =
+          tracer()
+              .getCallbackProvider(RequestContextSlot.APPSEC)
+              .getCallback(EVENTS.databaseSqlQuery());
+      if (sqlQueryCallback != null) {
+        RequestContext ctx = span.getRequestContext();
+        if (ctx != null) {
+          if (sql != null && !sql.isEmpty()) {
+            sqlQueryCallback.accept(ctx, sql);
+          }
+        }
+      }
+    }
     return span;
   }
