+++ dd-java-agent/instrumentation/jdbc/src/main/java/datadog/trace/instrumentation/jdbc/JDBCDecorator.java
@@ -1,18 +1,14 @@
 package datadog.trace.instrumentation.jdbc;
 
-import static datadog.trace.api.gateway.Events.EVENTS;
 import static datadog.trace.bootstrap.instrumentation.api.Tags.DB_OPERATION;
 import static datadog.trace.bootstrap.instrumentation.api.Tags.DB_SCHEMA;
 import static datadog.trace.bootstrap.instrumentation.api.Tags.DB_WAREHOUSE;
 
 import datadog.trace.api.Config;
 import datadog.trace.api.DDSpanId;
-import datadog.trace.api.gateway.RequestContext;
-import datadog.trace.api.gateway.RequestContextSlot;
 import datadog.trace.api.naming.SpanNaming;
 import datadog.trace.bootstrap.ContextStore;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
-import datadog.trace.bootstrap.instrumentation.api.AgentTracer;
 import datadog.trace.bootstrap.instrumentation.api.InternalSpanTypes;
 import datadog.trace.bootstrap.instrumentation.api.Tags;
 import datadog.trace.bootstrap.instrumentation.api.UTF8BytesString;
@@ -26,7 +22,6 @@ import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.HashSet;
 import java.util.Set;
-import java.util.function.BiConsumer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -55,11 +50,6 @@ public class JDBCDecorator extends DatabaseClientDecorator<DBInfo> {
       DBM_PROPAGATION_MODE.equals(DBM_PROPAGATION_MODE_FULL);
 
   private volatile boolean warnedAboutDBMPropagationMode = false; // to log a warning only once
-
-  // Extract this to allow for easier testing
-  protected AgentTracer.TracerAPI tracer() {
-    return AgentTracer.get();
-  }
 
   public static void logMissingQueryInfo(Statement statement) throws SQLException {
     if (log.isDebugEnabled()) {
@@ -113,6 +103,11 @@ public class JDBCDecorator extends DatabaseClientDecorator<DBInfo> {
   }
 
   @Override
+  protected String dbType(final DBInfo info) {
+    return info.getType();
+  }
+
+  @Override
   protected String dbUser(final DBInfo info) {
     return info.getUser();
   }
@@ -143,20 +138,6 @@ public class JDBCDecorator extends DatabaseClientDecorator<DBInfo> {
 
       setTagIfPresent(span, DB_WAREHOUSE, dbInfo.getWarehouse());
       setTagIfPresent(span, DB_SCHEMA, dbInfo.getSchema());
-
-      BiConsumer<RequestContext, String> connectDbCallback =
-          tracer()
-              .getCallbackProvider(RequestContextSlot.APPSEC)
-              .getCallback(EVENTS.databaseConnection());
-      if (connectDbCallback != null) {
-        RequestContext ctx = span.getRequestContext();
-        if (ctx != null) {
-          String dbType = dbInfo.getType();
-          if (dbType != null) {
-            connectDbCallback.accept(ctx, dbType);
-          }
-        }
-      }
     }
     return super.onConnection(span, dbInfo);
   }
@@ -227,27 +208,6 @@ public class JDBCDecorator extends DatabaseClientDecorator<DBInfo> {
       dbInfo = DBInfo.DEFAULT;
     }
     return dbInfo;
-  }
-
-  /**
-   * The method used to provide raw sql to prevent SQL-injection attacks SQL query should never be
-   * exposed because it may contain sensitive data.
-   */
-  public AgentSpan onStatementRaw(AgentSpan span, String sql) {
-    System.out.println("JDBCDecorator.onStatementRaw: " + sql);
-    BiConsumer<RequestContext, String> sqlQueryCallback =
-        tracer()
-            .getCallbackProvider(RequestContextSlot.APPSEC)
-            .getCallback(EVENTS.databaseSqlQuery());
-    if (sqlQueryCallback != null) {
-      RequestContext ctx = span.getRequestContext();
-      if (ctx != null) {
-        if (sql != null && !sql.isEmpty()) {
-          sqlQueryCallback.accept(ctx, sql);
-        }
-      }
-    }
-    return span;
   }
 
   public AgentSpan onStatement(AgentSpan span, DBQueryInfo dbQueryInfo) {
